+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changelog
 0.6.0 (2022-??-??)
 ------------------
 * Add `public_discovery` option to taxii2 config
+* Add support for publicly readable taxii 2 api roots
 
 0.5.0 (2022-05-24)
 ------------------